feat(hexagonal cycle 35): MailEventLoggingPort (port 15)
15ème port autonome. Signatures 100% primitives (&str), aucun type infrastructure. Couvre log_mail_event.
