+++ src/bin/email_api_dir/send_email_handler.rs
@@ -1,0 +1,87 @@
+use actix_web::{web, HttpResponse, Responder};
+use chrono::Utc;
+use serde::{Deserialize, Serialize};
+use uuid::Uuid;
+
+use simple_smtp_server::entities::Email;
+use simple_smtp_server::smtp_client::send_outgoing_email;
+
+use crate::DkimService;
+
+#[derive(Deserialize, Serialize, Debug, PartialEq)]
+pub struct EmailRequest {
+    pub from: String,
+    pub to: String,
+    pub subject: String,
+    pub body: String,
+}
+
+pub async fn send_email_handler(
+    email_req: web::Json<EmailRequest>,
+    dkim_service: web::Data<Box<dyn DkimService>>,
+) -> impl Responder {
+    println!("Received email request");
+
+    match dkim_service.sign_email(&email_req).await {
+        Ok(dkim_result) => {
+            println!("DKIM service returned success");
+            let message_id = dkim_result["messageId"].as_str().unwrap_or("");
+            match dkim_result["status"].as_str() {
+                Some("success") => {
+                    let sig = dkim_result["dkimSignature"]
+                        .as_str()
+                        .or_else(|| dkim_result["dkim_signature"].as_str())
+                        .unwrap_or("")
+                        .trim()
+                        .to_string();
+                    if sig.is_empty() {
+                        return HttpResponse::InternalServerError().json(serde_json::json!({
+                            "status": "error",
+                            "message": "DKIM service returned success without a DKIM signature",
+                        }));
+                    }
+
+                    let email = Email {
+                        id: Uuid::new_v4().to_string(),
+                        from: email_req.from.clone(),
+                        to: email_req.to.clone(),
+                        subject: email_req.subject.clone(),
+                        body: email_req.body.clone(),
+                        headers: vec![("DKIM-Signature".to_string(), sig.clone())],
+                        flags: vec![],
+                        sequence_number: 0,
+                        uid: 0,
+                        internal_date: Utc::now(),
+                        dkim_signature: Some(sig),
+                    };
+
+                    match send_outgoing_email(&email).await {
+                        Ok(_) => HttpResponse::Ok().json(serde_json::json!({
+                            "status": "success",
+                            "message": "Email signed and sent successfully",
+                            "messageId": message_id
+                        })),
+                        Err(e) => HttpResponse::InternalServerError().json(serde_json::json!({
+                            "status": "error",
+                            "message": format!("Failed to send email: {}", e)
+                        })),
+                    }
+                }
+                _ => {
+                    let error_message = dkim_result["message"].as_str().unwrap_or("Unknown error");
+                    HttpResponse::InternalServerError().json(serde_json::json!({
+                        "status": "error",
+                        "message": format!("Failed to sign email: {}", error_message)
+                    }))
+                }
+            }
+        }
+        Err(e) => {
+            eprintln!("Failed to call DKIM service: {}", e);
+            HttpResponse::InternalServerError().json(serde_json::json!({
+                "status": "error",
+                "message": "Failed to generate DKIM signature"
+            }))
+        }
+    }
+}
