+++ src/logic/traits.rs
@@ -221,3 +221,6 @@ pub use simple_smtp_domain::ports::CalendarPort;
 
 // Cycle 34 hexagonal — 14ème port : EmailStorePort.
 pub use simple_smtp_domain::ports::EmailStorePort;
+
+// Cycle 35 hexagonal — 15ème port : MailEventLoggingPort (journal mail events).
+pub use simple_smtp_domain::ports::MailEventLoggingPort;
